+++ test/bigtest/tests/QuickMarcEditor-test.js
@@ -14,10 +14,16 @@ const records = [
     content: {},
   },
   {
+    tag: '007',
+    content: {},
+  },
+  {
     tag: '008',
     content: {
       Type: 'a',
       BLvl: 'c',
+      ELvl: 'I',
+      Desc: 'i',
     },
   },
   {